{{cookiecutter.project_slug}}/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = "{{ cookiecutter.project_name }} - AI Development Conventions"
authors = [
    {name = "{{ cookiecutter.author_name }}", email = "{{ cookiecutter.author_email }}"},
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "PyYAML>=6.0",
    "textual>=0.47.0",
    "pydantic>=2.0",
]

[project.scripts]
ai-conventions = "ai_conventions.cli:main"

[tool.setuptools]
packages = ["ai_conventions", "ai_conventions.commands", "ai_conventions.providers"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-cookies>=0.6",
]